[refector] 구글 권장 아키텍처 적용
