+++ app/src/main/AndroidManifest.xml
@@ -16,10 +16,10 @@
         android:theme="@style/Theme.ZeroToMvvm"
         tools:targetApi="31">
         <activity
-            android:name=".view.detail.DetailActivity"
+            android:name=".ui.detail.DetailActivity"
             android:exported="false" />
         <activity
-            android:name=".view.main.MainActivity"
+            android:name=".ui.main.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
