--- app/src/main/java/com/kova700/zerotomvvm/data/source/pokemon/local/PokemonDao.kt
@@ -1,22 +1,0 @@
-package com.kova700.zerotomvvm.data.source.pokemon.local
-
-import androidx.room.Dao
-import androidx.room.Insert
-import androidx.room.OnConflictStrategy
-import androidx.room.Query
-import kotlinx.coroutines.flow.Flow
-
-@Dao
-interface PokemonDao {
-	@Insert(onConflict = OnConflictStrategy.IGNORE)
-	suspend fun insertPokemon(pokemon: PokemonEntity)
-
-	@Query("DELETE FROM PokemonEntity where num = :num")
-	suspend fun deletePokemon(num: Int)
-
-	@Query(
-		"SELECT * FROM PokemonEntity " +
-						"WHERE heart = :heartValue "
-	)
-	fun getWishPokemons(heartValue: Boolean = true): Flow<List<PokemonEntity>>
-}
