--- app/src/main/java/com/kova700/zerotomvvm/data/source/pokemon/remote/PokemonResponse.kt
@@ -1,13 +1,0 @@
-package com.kova700.zerotomvvm.data.source.pokemon.remote
-
-import com.kova700.zerotomvvm.data.source.pokemon.Pokemon
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class PokemonResponse(
-    @SerialName("count") val count: Int,
-    @SerialName("next") val next: String? = null,
-    @SerialName("previous") val previous: String? = null,
-    @SerialName("results") val results: List<Pokemon>
-)
