+++ app/src/main/java/com/kova700/zerotomvvm/domain/model/Pokemon.kt
@@ -1,0 +1,9 @@
+package com.kova700.zerotomvvm.domain.model
+
+data class Pokemon(
+	val num: Int,
+	val name: String,
+	val imageUrl: String,
+	val detailInfoUrl: String,
+	val heart :Boolean
+)
