+++ app/src/main/java/com/kova700/zerotomvvm/ui/main/MainActivity.kt
@@ -1,0 +1,41 @@
+package com.kova700.zerotomvvm.ui.main
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import androidx.navigation.NavController
+import androidx.navigation.fragment.NavHostFragment
+import androidx.navigation.fragment.findNavController
+import androidx.navigation.ui.setupWithNavController
+import com.kova700.zerotomvvm.R
+import com.kova700.zerotomvvm.databinding.ActivityMainBinding
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class MainActivity : AppCompatActivity() {
+    private lateinit var binding: ActivityMainBinding
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        binding = ActivityMainBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+        initNavigation()
+    }
+
+    private fun initNavigation() {
+        val navHostFragment =
+            supportFragmentManager.findFragmentById(R.id.container_main) as NavHostFragment
+        val navController = navHostFragment.findNavController()
+        initBottomNavigationView(navController)
+    }
+
+    private fun initBottomNavigationView(navController: NavController) {
+        binding.bnvMain.apply {
+            setupWithNavController(navController)
+        }
+    }
+
+    companion object {
+        const val TO_DETAIL_SELECTED_ITEM_INDEX_EXTRA = "TO_DETAIL_SELECTED_ITEM_INDEX_EXTRA"
+        const val TO_DETAIL_SELECTED_ITEM_STATE_EXTRA = "TO_DETAIL_SELECTED_ITEM_STATE_EXTRA"
+    }
+}
