+++ app/src/main/java/com/kova700/zerotomvvm/ui/main/model/SelectedItemState.kt
@@ -1,0 +1,5 @@
+package com.kova700.zerotomvvm.ui.main.model
+
+enum class SelectedItemState {
+	DEFAULT, WISH
+}
