+++ app/src/main/res/layout/activity_detail.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.detail.DetailActivity">
+    tools:context=".ui.detail.DetailActivity">
 
     <com.google.android.material.imageview.ShapeableImageView
         android:id="@+id/siv_detail_activity"
