+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,13 +7,13 @@
 
     <fragment
         android:id="@+id/homeFragment"
-        android:name="com.kova700.zerotomvvm.view.main.home.HomeFragment"
+        android:name="com.kova700.zerotomvvm.ui.main.home.HomeFragment"
         android:label="fragment_home"
         tools:layout="@layout/fragment_home">
     </fragment>
     <fragment
         android:id="@+id/wishFragment"
-        android:name="com.kova700.zerotomvvm.view.main.wish.WishFragment"
+        android:name="com.kova700.zerotomvvm.ui.main.wish.WishFragment"
         android:label="fragment_wish"
         tools:layout="@layout/fragment_wish">
     </fragment>
